requirements.txt
appnope==0.1.0
backcall==0.1.0
cffi==1.11.5
cycler==0.10.0
data==0.4
decorator==4.3.0
funcsigs==1.0.2
future==0.16.0
ipython-genutils==0.2.0
jedi==0.12.0
kiwisolver==1.0.1
latex==0.7.0
matplotlib==2.2.2
numpy==1.14.2
pandas==0.22.0
parso==0.2.0
path.py==7.7.1
pexpect==4.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.2
python-igraph==0.7.1.post6
pytz==2018.4
scipy==1.0.1
shutilwhich==1.1.0
simplegeneric==0.8.1
six==1.11.0
tempdir==0.7.1
traitlets==4.3.2
wcwidth==0.1.7
